tried to tune the FPS (frame rate)
- in FPS::calc, rate renamed _period and now in the class declaration.
commented out the stats stuff ... does not seem to be used.
In FPS::delay, added an other nanosleep function for the last 100 us
+ a viariable for fine tuning. Seems to work not to bad on my machine,
but needs to be tested on others.

- In VideoLayer, deleted some printed information.

+++ src/include/context.h
@@ -164,7 +164,7 @@ class Context {
   /* Set the interval (in frames) after
      the fps counter is updated */
   FPS fps;
-  float fps_speed;
+  double fps_speed;
 
   bool clear_all;
   bool start_running;
